requirements.txt
streamlit
flask
tensorflow
numpy
Pillow